Move IDS tests to main tests class.
Add comment of how to run specific algorithm tests.
